requirements.txt
torch
torchvision
torchaudio
numpy
PyYAML
Pillow
tqdm
setuptools